Added basic support for PSR-5 <> type notation.
Since the notation is not final, all this patch
does is accept the values and provide the information
using the same object already available.

array<int> will be parsed as if there was int[].
array<int, int> will set the array key type but
not do muche extra.

Object<this,will,be,ignored> will return an Object_
type of the right type, no way is provided to get
the inner type at this moment.