Build a Lexer based on Doctrine/Lexer
As an experiment, I have built a lexer for the TypeResolver using
Doctrine's Lexer library. This Lexer will create tokens out of a sting
representing a type.

Net step in this experiment is to build a parser, and replace the
internals of the TypeResolver class to use this lexer and parser. The
goal is to come to a BC-compatible solution, which can also be used to
make benchmarks to see which version runs faster _and_ is more readable.