Phive: upgrade used version of PHPUnit
The tests were not actually running on PHP 8, nor PHP 8.1, due to PHPUnit 8.4 being used, which is not compatible with PHP 8.0+.
The first compatible version in the 8.x range is 8.5 and on the 9.x. range, 9.3.
As for PHP 8.1, there is currently no guarantee yet, but your best bet is using the latest release of PHPUnit.

The `push.yml` script already contains code to _downgrade_ the PHPUnit version used by Phive for PHP 7.2 back to PHPUnit 8, so it should be safe to upgrade the PHPUnit version for PHP 7.3 and above to PHPUnit 9.5.x.

The only concession which needs to be made is for the `phpunit-with-coverage` to be run on PHP 7.3 (or to also downgrade to PHPUnit 8).