Remove legacy nullable format
Compound types containg `?` nullables are deprecated in php, and
not supported by phpstan. In v1 we had a work around for this but
this was slow and fragile. The nullable format has been deprecated
since 2023 and is now removed. We still produce a deprecation error
to warn people about invalid type handling.

fixes #184